Viper + Biomass Gun Update

+++ VVUP.CustomItems/Items/Firearms/Scp2818.cs
@@ -23,11 +23,17 @@ namespace VVUP.CustomItems.Items.Firearms
         public override float Damage { get; set; } = 1000;
         public override byte ClipSize { get; set; } = 1;
 
-        [Description("Whether or not the weapon should despawn itself after it's been used.")]
+        [Description("Whether or not the weapon should despawn itself after it was shot and did not miss.")]
         public bool DespawnAfterUse { get; set; } = true;
+        [Description("Whether or not the weapon should despawn itself after it was shot and missed.")]
+        public bool DespawnAfterMiss { get; set; } = true;
+        [Description("Should the player instantly die after missing a shot with this weapon?")]
+        public bool KillAfterMiss { get; set; } = true;
+        [Description("Should it check if wherever the player shot was the floor? Note: Disabling this will allow players to get into spots they shouldn't be able to be in.")]
+        public bool UseRaycastPosition { get; set; } = true;
 
-        public string DeathReasonUser { get; set; } = "Vaporized by becoming a bullet";
-        public string DeathReasonTarget { get; set; } = "Vaporized by a human bullet";
+        public string DeathReasonUser { get; set; } = "Vaporized by becoming a bullet.";
+        public string DeathReasonTarget { get; set; } = "Vaporized by a human bullet.";
 
         public override SpawnProperties SpawnProperties { get; set; } = new()
         {
@@ -49,10 +55,10 @@ namespace VVUP.CustomItems.Items.Firearms
 
         protected override void OnShot(ShotEventArgs ev)
         {
+            bool miss = false;
             if (ev.Target == null)
             {
-                Log.Debug($"VVUP Custom Items: SCP-2818, {ev.Player.Nickname} fired and missed a target, teleporting them to bullet impact location ({ev.Position}");
-                ev.Player.Position = ev.Position;
+                miss = true;
             }
             else
             {
@@ -71,16 +77,57 @@ namespace VVUP.CustomItems.Items.Firearms
                 }
             }
 
-            if (DespawnAfterUse)
-            {
-                Log.Debug($"VVUP Custom Items: SCP-2818, Despawn After Use is true, removing SCP-2818 from {ev.Player.Nickname}'s inventory");
-                ev.Player.RemoveItem(ev.Item);
-            }
+
 
             Timing.CallDelayed(0.1f, () =>
             {
-                Log.Debug($"VVUP Custom Items: SCP-2818, Killing {ev.Player.Nickname}");
-                ev.Player.Kill(DeathReasonUser);
+                if (miss)
+                {
+                    Log.Debug($"VVUP Custom Items: SCP-2818, {ev.Player.Nickname} fired and missed a target, teleporting them to bullet impact location ({ev.RaycastHit.point}");
+
+                    if (DespawnAfterMiss)
+                    {
+                        ev.Player.RemoveHeldItem();
+                    } else
+                    {
+                        ev.Player.DropHeldItem();
+                    }
+
+                    ev.CanSpawnImpactEffects = false;
+
+                    if (UseRaycastPosition)
+                    {
+                        if (ev.RaycastHit.normal == new UnityEngine.Vector3(0, 1, 0))
+                        {
+                            ev.Player.Position = ev.RaycastHit.point + new UnityEngine.Vector3(0, 1, 0);
+                        }
+                    } else
+                    {
+                        ev.Player.Position = ev.Position;
+                    }
+
+                    Timing.CallDelayed(0.1f, () =>
+                    {
+                        if (KillAfterMiss)
+                        {
+                            ev.Player.Kill(DeathReasonUser);
+                        } else
+                        {
+                            ev.Player.Hurt(ev.Distance, "Body is badly mutilated.");
+                        }
+                    });
+                }
+                else
+                {
+                    if (DespawnAfterUse)
+                    {
+                        Log.Debug($"VVUP Custom Items: SCP-2818, Despawn After Use is true, removing SCP-2818 from {ev.Player.Nickname}'s inventory");
+                        ev.Player.RemoveItem(ev.Item);
+                    }
+
+                    Log.Debug($"VVUP Custom Items: SCP-2818, Killing {ev.Player.Nickname}");
+                    ev.Player.Kill(DeathReasonUser);
+                }
             });
         }
     }
